requirements.txt
aiohttp
dnspython
argparse
httpx
